+++ CHANGELOG.md
@@ -7,7 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-- Nothing worth mentioning yet.
+### Added
+
+- New settings category: `Content Feed`
+  - New setting: `Disable Shorts`
 
 ## [0.38.2] - 2025-05-22